agents/backend-engineer/daily_intake_automation/requirements.txt
pydantic>=2.0.0
pydantic-settings>=2.0.0
psycopg2-binary>=2.9.9
requests>=2.31.0
python-dotenv>=1.0.0
structlog>=23.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
httpx>=0.25.0
